added scrollview to homepage
