+++ frontend/app/components/DisplayButtons.js
@@ -76,11 +76,12 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: 'white',
-    padding: 10,
   },
   actionButtonIcon: {
     fontSize: 20,
     height: 22,
     color: 'white',
+    bottom: '0%',
+    right: '0%',
   },
 });
